Clean up commonWaitFor implementation
